Ensure uv pip freeze is colorless
The 'uv pip freeze' command should automatically determine whether or
not to provide coloring if the output is going to a terminal or TTY with
support. This is not happening in this case.

Ensuring color is always disabled for this command is a good first step
to work around this issue.